refactor(category): separate category selectors
